3.1.8. Move constans to Interface
