+++ src/main/java/edu/kis/vh/nursery/stack/Stack.java
@@ -2,6 +2,8 @@ package edu.kis.vh.nursery.stack;
 
 public interface Stack {
 
+    int EMPTY_STACK_VALUE = -1;
+
     void push(int newNodeValue);
 
     boolean isEmpty();
